changed character IDs to be consistent with teams

+++ Common/Contracts/DTO/ServerComands/ProjectileNotification.cs
@@ -7,5 +7,5 @@ namespace OpenTrenches.Common.Contracts.DTO.ServerComands;
 public record class ProjectileNotificationCommand(
     [property: Key(0)] Vector2 Start,
     [property: Key(1)] Vector2 End,
-    [property: Key(2)] ushort Character) : AbstractCommandDTO
+    [property: Key(2)] int Character) : AbstractCommandDTO
 {}
