+++ Common/Contracts/DTO/ServerComands/ReloadNotificationCommand.cs
@@ -4,6 +4,6 @@ namespace OpenTrenches.Common.Contracts.DTO.ServerComands;
 
 [MessagePackObject]
 public record class ReloadNotificationCommand(
-    [property: Key(0)] ushort Character
+    [property: Key(0)] int Character
 ) : AbstractCommandDTO
 {}
