+++ Core/Scripting/IIdObject.cs
@@ -4,5 +4,5 @@ namespace OpenTrenches.Core.Scripting;
 
 public interface IIdObject
 {
-    public ushort ID { get; }
+    public int ID { get; }
 }
